weapon

+++ DGP.Genshin/Controls/Converters/StarToImageSourceConverter.cs
@@ -9,7 +9,7 @@ namespace DGP.Genshin.Controls.Converters
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return Application.Current.FindResource((int)value + "star");
+            return Application.Current.FindResource(value + "star");
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) => throw new NotImplementedException();
